move code to the pallet
